Fix disk detection to work when used on a node with an alias
As the webui uses the alias in preference to the name, it will also pass
the alias as argument to the controller when fetching the list of disks.

So the controller should support finding a node by alias and by name.

Also, just be nice and don't crash if we don't find a node :-)

https://bugzilla.novell.com/show_bug.cgi?id=777897