In non-SSL mode with no public name, use IP address for public endpoint
After some discussion, we prefer to use the IP address if SSL is not
used and if there is no public name set for the node, instead of using
'public.'+fqdn.

This decreases the risk of issues in case DNS is not properly setup.

Part of https://bugzilla.novell.com/show_bug.cgi?id=777523