Fix javascript listing the volumes
Since 69346e11 in barclamp-crowbar, the text used when listing nodes for
each role is "<shortname> <fullname>" instead of "<shortname>". This
broke the javascript function using this text to query the list of disk
for the node.

So just remove everything after ' ' in the text when passing it as an
argument for the query.