Do not mark new options as required
This creates an issue for existing proposals, since the JSON data will
only have the new options if they get changed; most users trying to save
will face an error because of the schema.

Having them not required is okay since the cookbooks is using a sane
default for the values.